user deleting for some reason

+++ src/Adopt.js
@@ -5,40 +5,34 @@ import Dogs from "./Dogs/Dogs";
 import userService from "./services/userService";
 
 export default class Home extends Component {
-  state = { status: {}, count: 0 };
 
   static contextType = Context;
-  //static interval =
-  //setInterval(() => userService.deleteCurrent, 5000)%%
+  
   componentDidMount() {
-    this.interval = setInterval(() => {
-      this.shuffleUser();
-    }, 10000);
-    //this.displayQueue();%%
+    this.context.loadUsers()
+    if(this.context.user !== this.context.users[0]){
+        this.interval = setInterval(() => {
+        this.shuffleUser();
+        this.context.addRandom('random')
+        }, 3000);
+    }
   }
-  //check if I'm on my last user? condition
-  shuffleUser = () => {
-    this.context.removeUser(); //updates view
-    const count = this.state.count++;
-    this.setState({ status: {}, count: count });
-    userService.deleteCurrent();
-    this.context.removeUser();
-  };
-  // The currentUser (users[0]) should be deleted every five seconds. The currentUser is located
-  // in state in the app module, and available through Context. Once our logged-in user is at users[0],
-  // the timeout function should stop, allowing the user to choose a pet.
-  // displayQueue = () => {%%
-  //   for (let i = 0; i < this.context.users.length; i++) {%%
-  //     if (this.context.user !== this.context.currentUser) {%%
-  // setInterval(() => userService.deleteCurrent, 5000);%%
-  //     }
-  //   }
-  //   //call a method to force a state update
-  // };
 
-  addToQueue = () => {
-    // There should be a new random user added to the end of the queue every few seconds.
+  shuffleUser = () => {
+    if (this.context.users.length > 1) {
+        userService.deleteCurrent()
+        .then(this.context.removeUser());
+    }
   };
+
+//   addToQueue = () => {
+//     while(this.context.users.length < 5){
+//         setInterval(() => {
+//             userService.postUser('random')
+//             this.context.loadUsers()
+//         }, 5000);
+//     }
+//   };
 
   render() {
     if (this.context.isLoading) {
